% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee_ellipse_area.R
\name{apogee_ellipse_area}
\alias{apogee_ellipse_area}
\title{Calculation of the field of view of Apogee® radiometers}
\usage{
apogee_ellipse_area(
  ang_sfn,
  ang_half = NULL,
  h_i,
  from_model = FALSE,
  rm_model = NULL
)
}
\arguments{
\item{ang_sfn}{Angle to surface normal as deviation between vertical axis and line of sight (numeric). Must be in degrees}

\item{ang_half}{Half angle to surface normal as deviation between vertical axis and line of sight (numeric). Must be in degrees}

\item{h_i}{Height of the instrument (numeric). Must be in meter}

\item{from_model}{States whether ang_half is derived from 'rm_model' (boolean). If TRUE, 'rm_model' must be given. Default = FALSE}

\item{rm_model}{Radiometer type if the model-specific 'ang_half' is unknown (character).
Allowed inputs are: "SI-111", "SI-121" or "SI-131"}
}
\value{
Either a spatial raster or a numeric value, depending on the function input
}
\description{
`apogee_ellipse_area` calculates the land surface temperature from radiance temperature
}
\author{
Lukas Regensburger
}
