% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_sensor_drift.R
\name{corr_sensor_drift}
\alias{corr_sensor_drift}
\title{Correction of temperature values acquired by an unstable thermal camera
Adds the amount of necessary correction and the corrected temperature as new_columns to an existing dataframe}
\usage{
corr_sensor_drift(
  df,
  col_tsens,
  col_traw,
  col_time,
  method = "sensor_only",
  fit = "spline",
  degf = 2,
  threshold = 0.3,
  datetime_format
)
}
\arguments{
\item{df}{Input table (dataframe). Must at least include columns for 'col_tsens', 'col_traw' and 'col_time'}

\item{col_tsens}{Name of the column featuring the sensor temperature values (character). Values must be in degrees Celsius}

\item{col_traw}{Name of the column featuring the recorded temperature values, usually mean values of single images (character).
Values must be in degrees Celsius.}

\item{col_time}{Name of the column featuring acquisition time stamps (character).
Must be in a recognized datetime format, which is specified by 'datetime_format'}

\item{method}{Method used to calculate deviations and corrected temperatures (character). Either "sensor_only" or "lst_dev", see also the README-file}

\item{fit}{Mathematical method used for modelling variable relations (character). Either "polynomial" or "spline"; default is "spline"}

\item{degf}{Degrees of freedom for function fitting (numeric). Default is 2. Increase with care to prevent overfitting}

\item{threshold}{Margin from minimum 'tsens' to define interval where 'tsens' is stable (numeric). Default is 0.3}

\item{datetime_format}{Describes format of 'col_time' to the algorithm (character).
Must refer to a recognized datetime format. Common example would be "%Y-%m-%d %H:%M:%S"}
}
\value{
Input table with amount of correction and corrected temperature added as new columns
}
\description{
`corr_sensor_drift` provides two correction methods to compensate unstable thermal cameras
}
\author{
Lukas Regensburger
}
