% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2lst.R
\name{rad2lst}
\alias{rad2lst}
\title{Calculation of the land surface temperature from radiance temperature.
Additionally, the ambient background temperature can be returned or, if existing, used for calculation.}
\usage{
rad2lst(
  trad,
  emi,
  lw_in,
  use_tbg = FALSE,
  return_tbg = FALSE,
  tbg = NULL,
  rast_tol_fact = 0.1
)
}
\arguments{
\item{trad}{Radiance temperature as single number (numeric) or raster image (SpatRaster)}

\item{emi}{Surface emissivity as single number (numeric) or raster image (SpatRaster). Must be between 0 and 1.}

\item{lw_in}{Incoming longwave radiation as single number (numeric) or raster image (SpatRaster). Must be > 0.}

\item{use_tbg}{States whether background temperature instead of lw_in is used (boolean). If TRUE, 'return_tbg' must be FALSE.}

\item{return_tbg}{States whether'tbg' is returned as well when using lw_in (boolean). Only possible if 'use_tbg' is FALSE.}

\item{tbg}{Background-, also referred to as ambient temperature, as single number (numeric) or raster image (SpatRaster)}

\item{rast_tol_fact}{Applied when computing a minimum tolerance for raster comparison (numeric). Only needed if more than 1 input is a raster.
Approximate values for orientation are < 0.5 when coordinates are in degree and < 0.001 when in geographical units.}
}
\value{
Either a spatial raster or a numeric value, depending on the function input
}
\description{
`rad2lst` calculates the land surface temperature from radiance temperature
}
\author{
Lukas Regensburger
}
